- Move FlagWarPlace listener to FlagWarBlockListener.
