+++ resources/ChangeLog.txt
@@ -5327,7 +5327,7 @@ v0.92.0.11:
     - Closes #4100.
   - Switch Explosive Fireworks to be handled like PVP damage when it is dealt to players, instead of Explosive Damage.
     - Makes it possible to use them on crossbows in a pvp setting without needing explosions forced on.
-0.96.3.8
+0.96.3.8:
   - Update bstats metrics.
   - Fix javadoc.
   - Add getBlock() to TownyBuildEvent, TownyDestroyEvent and TownySwitchEvent.
@@ -5340,3 +5340,7 @@ v0.92.0.11:
     - Move new town method to TownyUniverse.
     - Create a InvalidNameException that is a TownyException.
 	- Create register / unregister methods for towns.
+0.6.3.9:
+  - Fix nation neutral messages, courtesy of silverwolfg11 with PR #4468.
+    - Closes #4467.
+  - Small Clean-up, Sticky pistons cannot move immutable FlagWar blocks (e.g. a flag beacon), courtesy of TheFlagCourier with PR #4469.
